[tools] perf pmu: Fix check for unset alias->unit array
Message-id: <20170807084247.7531-66-jolsa@redhat.com>
Patchwork-id: 182002
O-Subject: [PATCH RHEL7.5 65/94] perf pmu: Fix check for unset alias->unit array
Bugzilla: 1478485
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1478485
https://bugzilla.redhat.com/show_bug.cgi?id=1478485

upstream
========
commit b30a7d1fc96d60fe2c02e375a56a046385701d17
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Feb 15 10:06:20 2017 -0300

description
===========
The alias->unit field is an array, so to check that it is not set we
should see if it is an empty string, i.e. alias->unit[0], instead of
checking alias->unit != NULL, as this will _always_ evaluate to 'true'.

Pointed out by clang.

Signed-off-by: Rafael Aquini <aquini@redhat.com>