[tools] scripts/spelling.txt: add "an user" pattern and fix typo instances
Message-id: <20170807084247.7531-71-jolsa@redhat.com>
Patchwork-id: 182009
O-Subject: [PATCH RHEL7.5 70/94] scripts/spelling.txt: add "an user" pattern and fix typo instances
Bugzilla: 1478485
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1478485
https://bugzilla.redhat.com/show_bug.cgi?id=1478485

upstream
========
commit 9332ef9dbd172d4ab0a0141df7cb21c696a5ce96
Author: Masahiro Yamada <yamada.masahiro@socionext.com>
Date: Mon Feb 27 14:28:47 2017 -0800

description
===========
Fix typos and add the following to the scripts/spelling.txt:

  an user||a user
  an userspace||a userspace

I also added "userspace" to the list since it is a common word in Linux.
I found some instances for "an userfaultfd", but I did not add it to the
list.  I felt it is endless to find words that start with "user" such as
"userland" etc., so must draw a line somewhere.

RHEL7
=====
Taking just perf related hunk.

Signed-off-by: Rafael Aquini <aquini@redhat.com>