[tools] perf symbols: Fix symbols__fixup_end heuristic for corner cases
Message-id: <20170807084247.7531-95-jolsa@redhat.com>
Patchwork-id: 182028
O-Subject: [PATCH RHEL7.5 94/94] perf symbols: Fix symbols__fixup_end heuristic for corner cases
Bugzilla: 1478485
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1478485
https://bugzilla.redhat.com/show_bug.cgi?id=1478485

upstream
========
commit e7ede72a6d40cb3a30c087142d79381ca8a31dab
Author: Daniel Borkmann <daniel@iogearbox.net>
Date: Wed Mar 15 22:53:37 2017 +0100

description
===========
The current symbols__fixup_end() heuristic for the last entry in the rb
tree is suboptimal as it leads to not being able to recognize the symbol
in the call graph in a couple of corner cases, for example:

 i) If the symbol has a start address (f.e. exposed via kallsyms)
    that is at a page boundary, then the roundup(curr->start, 4096)
    for the last entry will result in curr->start == curr->end with
    a symbol length of zero.

ii) If the symbol has a start address that is shortly before a page
    boundary, then also here, curr->end - curr->start will just be
    very few bytes, where it's unrealistic that we could perform a
    match against.

Instead, change the heuristic to roundup(curr->start, 4096) + 4096, so
that we can catch such corner cases and have a better chance to find
that specific symbol. It's still just best effort as the real end of the
symbol is unknown to us (and could even be at a larger offset than the
current range), but better than the current situation.

Alexei reported that he recently run into case i) with a JITed eBPF
program (these are all page aligned) as the last symbol which wasn't
properly shown in the call graph (while other eBPF program symbols in
the rb tree were displayed correctly). Since this is a generic issue,
lets try to improve the heuristic a bit.

Fixes: 2e538c4a1847 ("perf tools: Improve kernel/modules symbol lookup")
Signed-off-by: Rafael Aquini <aquini@redhat.com>