[fs] gfs2: Introduce helpers for setting and clearing gl_object
Message-id: <20170809170306.8691-2-rpeterso@redhat.com>
Patchwork-id: 182502
O-Subject: [RHEL7.5 fs PATCH v4 1/4] GFS2: Introduce helpers for setting and clearing gl_object
Bugzilla: 1471151
RH-Acked-by: Andreas Grunbacher <agruenba@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

This patch introduces new helper functions in glock.h that set
and clear gl_object, with an added integrity check to ensure
the previous values are not overwritten. Helper function
glock_set_object was introduced by first the upstream patch,
but for a different reason unrelated to this bug.

So I've excerpted that portion of the first patch. The second
patch implements the reciprocal function glock_clear_object,
but it does not contain the spin_locks from the upstream
version. Again, this is due to the requirements of the first
patch which are not carried over into rhel7.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1471151

This is a RHEL7 port based on these two upstream patches:
https://git.kernel.org/pub/scm/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=4fd1a5795214bc6405f14691c1344ae8c3f17215
https://git.kernel.org/pub/scm/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=df3d87bde121213560fde0edb71bc46f0f75692c
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>