[fs] gfs2: Set gl_object in inode lookup only after block type check
Message-id: <20170809170306.8691-3-rpeterso@redhat.com>
Patchwork-id: 182503
O-Subject: [RHEL7.5 fs PATCH v4 2/4] GFS2: Set gl_object in inode lookup only after block type check
Bugzilla: 1471151
RH-Acked-by: Andreas Grunbacher <agruenba@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

Before this patch, the inode glock's gl_object was set after a
reference was acquired, but before the block type was verified.
In cases where the block was unlinked, then freed and reused on
another node, a residual delete callback (delete_work) would try
to look up the inode, eventually failing the block check, but
only after it overwrites gl_object with a pointer to the wrong
inode. This patch moves the assignment of gl_object after the
block check so it won't be improperly overwritten. This ensures
proper locking is maintained when the gl_object is set.

Likewise, at the end of the function, gfs2_inode_lookup was
clearing gl_object after it unlocked the glock, which meant
another process might free the glock in the meantime. This
patch guards against that case.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
Reviewed-by: Andreas Gruenbacher <agruenba@redhat.com>

rhbz#1471151

This is a RHEL7 port of the following upstream patch:
https://git.kernel.org/pub/scm/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=4d7c18c7df89ef549f2de79b0faf873b49dea57a
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>