[fs] gfs2: Clear gl_object if gfs2_create_inode fails
Message-id: <20170809170306.8691-4-rpeterso@redhat.com>
Patchwork-id: 182501
O-Subject: [RHEL7.5 fs PATCH v4 3/4] GFS2: Clear gl_object if gfs2_create_inode fails
Bugzilla: 1471151
RH-Acked-by: Andreas Grunbacher <agruenba@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

This patch fixes a bug in an error path in which the gl_object
value was left pointing at freed memory.

If function gfs2_create_inode fails after the inode has been
created (for example, if the inode_refresh fails for some reason)
the function was setting gl_object but never clearing it again.
The glocks are left pointing to a freed inode. This patch adds
the calls to clear gl_object in the appropriate error paths.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1471151

This is a RHEL7 port of the following upstream patch:
https://git.kernel.org/pub/scm/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=9c1b28081f43c0f14ccbcad02a6e0f227c072da2
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>