[fs] gfs2: Clear gl_object when deleting an inode in gfs2_delete_inode
Message-id: <20170809170306.8691-5-rpeterso@redhat.com>
Patchwork-id: 182504
O-Subject: [RHEL7.5 fs PATCH v4 4/4] GFS2: Clear gl_object when deleting an inode in gfs2_delete_inode
Bugzilla: 1471151
RH-Acked-by: Andreas Grunbacher <agruenba@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

This patch fixes a bug in an error path in which the gl_object value
was left pointing at freed memory.

This patch adds some calls to clear gl_object in function
gfs2_delete_inode. Since we are deleting the inode, and the glock
typically outlives the inode in core, we must clear gl_object
so subsequent use of the glock (e.g. for a new inode in its place)
will not have the old pointer sitting there. In error cases we
need to tidy up after ourselves. In non-error cases, we need to
clear gl_object before we set the block free in the bitmap so
residuals aren't left for potential inode creators.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1471151

This is a RHEL7 port of the following upstream patch:
https://git.kernel.org/pub/scm/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2?h=for-next&id=240c6235dfe4fe46f2fd0f2b3c15b3a22100a75e
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>