[fs] sunrpc: Handle EADDRNOTAVAIL on connection failures
Message-id: <1502400187-7158-1-git-send-email-dwysocha@redhat.com>
Patchwork-id: 182596
O-Subject: [RHEL7.5 fs PATCH BZ 1479043] SUNRPC: Handle EADDRNOTAVAIL on connection failures
Bugzilla: 1479043
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1479043
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13794216
Upstream Status: Merged
Testing: 100% reproducible with testcase on the bug

So far one customer has reported this problem on RHEL7.3 and we've had multiple
customers report this on RHEL6.  This patch was previously posted for RHEL6.10
and now I'm posting for RHEL7.5.

The customer that hit this on a RHEL7.3 kernel reported their NFS share went
unresponsive for 60 seconds after a connection dropped.  We checked and they
had TCP timestamps disabled and were using NFSv3 so will hit this problem.

This patch fixes a sunrpc regression introduced in RHEL7.3 by the patchset for
bug 1270038 as a result of re-using the same port when TCP timestamps are disabled.
When TCP timestamps are disabled, during the TIME_WAIT state of a TCP connection
kernel_connect returns EADDRNOTAVAIL for the duration of TIME_WAIT if you try to
connect a socket with the same source port.  Since the SO_REUSEPORT patches
went in (bz 1270038), this means NFS client cannot re-use the port during TIME_WAIT,
which lasts for 60 seconds on any TCP connection.  The result is an NFS share can
be unavailable for up to 60 seconds in these environments.  Some environments like
NAT require TCP timestamps to be disabled, and these environments will see this
issue show up.  The patch below fixes this by picking another port, which is the
same way upstream handles this case.

We have reproduced this problem and shown the patch fixes the issue and returns
RHEL7 to the state it was in for RHEL7.2 and also how upstream behave.
The patch was a clean backport to RHEL7.

From: Trond Myklebust <trond.myklebust@primarydata.com>

If the connect attempt immediately fails with an EADDRNOTAVAIL error, then
that means our choice of source port number was bad.
This error is expected when we set the SO_REUSEPORT socket option and we
have 2 sockets sharing the same source and destination address and port
combinations.

Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Fixes: 402e23b4ed9ed ("SUNRPC: Fix stupid typo in xs_sock_set_reuseport")
Cc: stable@vger.kernel.org # v4.0+
(cherry picked from commit 1f4c17a03ba7f430d63dba8c8e08ff1e2712581d)
Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>