[fs] nfsv4: add flock_owner to open context
Message-id: <6a4462e27f6826b05baa43bd365265bc02935027.1503324952.git.bcodding@redhat.com>
Patchwork-id: 182958
O-Subject: [RHEL7 fs PATCH 2/7] NFSv4: add flock_owner to open context
Bugzilla: 1476826
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Scott Mayhew <smayhew@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: NeilBrown <neilb@suse.com>

An open file description (struct file) in a given process can be
associated with two different lock owners.

It can have a Posix lock owner which will be different in each process
that has a fd on the file.
It can have a Flock owner which will be the same in all processes.

When searching for a lock stateid to use, we need to consider both of these
owners

So add a new "flock_owner" to the "nfs_open_context" (of which there
is one for each open file description).

This flock_owner does not need to be reference-counted as there is a
1-1 relation between 'struct file' and nfs open contexts,
and it will never be part of a list of contexts.  So there is no need
for a 'flock_context' - just the owner is enough.

The io_count included in the (Posix) lock_context provides no
guarantee that all read-aheads that could use the state have
completed, so not supporting it for flock locks in not a serious
problem.  Synchronization between flock and read-ahead can be added
later if needed.

When creating an open_context for a non-openning create call, we don't have
a 'struct file' to pass in, so the lock context gets initialized with
a NULL owner, but this will never be used.

The flock_owner is not used at all in this patch, that will come later.

Acked-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: NeilBrown <neilb@suse.com>
Reviewed-by: Jeff Layton <jlayton@redhat.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
(cherry picked from commit 532d4def2f95623a9b8b2cef7723e14521377911)
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1476826
Signed-off-by: Rafael Aquini <aquini@redhat.com>