[fs] gfs2: Withdraw for IO errors writing to the journal or statfs
Message-id: <856866647.919598.1503423735381.JavaMail.zimbra@redhat.com>
Patchwork-id: 183110
O-Subject: [RHEL7.5 fs PATCH] GFS2: Withdraw for IO errors writing to the journal or statfs
Bugzilla: 1429547
RH-Acked-by: Abhijith Das <adas@redhat.com>
RH-Acked-by: Andrew Price <anprice@redhat.com>

Hi,

Before this patch, if GFS2 encountered IO errors while writing to
the journal, it would not report the problem, so they would go
unnoticed, sometimes for many hours. Sometimes this would only be
noticed later, when recovery tried to do journal replay and failed
due to invalid metadata at the blocks that resulted in IO errors.

This patch makes GFS2's log daemon check for IO errors. If it
encounters one, it withdraws from the file system and reports
why in dmesg. A similar action is taken when IO errors occur when
writing to the system statfs file.

These errors are also reported back to any callers of fsync, since
that requires the journal to be flushed. Therefore, any IO errors
that would previously go unnoticed are now noticed and the file
system is withdrawn as early as possible, thus preventing further
file system damage.

Signed-off-by: Bob Peterson <rpeterso@redhat.com>

rhbz#1429547

This is a RHEL7 port of the following upstream patch:
https://git.kernel.org/pub/scm/linux/kernel/git/gfs2/linux-gfs2.git/commit/fs/gfs2/incore.h?h=for-next&id=9bfef7554e20d7bac00cddd976f5abb5b2d1b198

The patch was built into several kernels, including this one:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=13899788

It was unit tested on system gfs-i24c-01.mpc.lab.eng.bos.redhat.com
using dm-flakey and the scenario describe in the bugzilla, comment 8.
It was further tested with sas calibration on the afc cluster.
---

Signed-off-by: Rafael Aquini <aquini@redhat.com>