[fs] cifs: Add a modified cifs_read_page_from_socket()
Message-id: <20170825113940.27960-24-sprabhu@redhat.com>
Patchwork-id: 183270
O-Subject: [RHEL7.5 PATCH fs 23/28] [bz1429710]cifs: Add a modified cifs_read_page_from_socket()
Bugzilla: 1429710
RH-Acked-by: Jeff Layton <jlayton@redhat.com>
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Leif Sahlberg <lsahlber@redhat.com>

Fixes: bz 1429710

This is needed by the next patch in the series in the function
read_data_into_pages().

The original function upstream is added by the commit 71335664c38f
("cifs: don't bother with kmap on read_pages side") which uses the bvec
iter which is not available in RHEL 7.

Signed-off-by: Sachin Prabhu <sprabhu@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>