[x86] perf/x86/intel/pt: Allow the disabling of branch tracing
Message-id: <20170830160540.26618-6-jolsa@redhat.com>
Patchwork-id: 183417
O-Subject: [PATCH RHEL7.5 BZ1480522 005/248] perf/x86/intel/pt: Allow the disabling of branch tracing
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit d35869ba348d3f1ff3e6d8214fe0f674bb0e404e
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Mon Feb 6 16:41:40 2017 +0200

description
===========
Now that Intel PT supports more types of trace content than just branch
tracing, it may be useful to allow the user to disable branch tracing
when it is not needed.

The special case is BDW, where not setting BranchEn is not supported.

This is slightly trickier than necessary, because up to this moment
the driver has been setting BranchEn automatically and the userspace
assumes as much. Instead of reversing the semantics of BranchEn, we
introduce a 'passthrough' bit, which will forego the default and allow
the user to set BranchEn to their heart's content.

Signed-off-by: Rafael Aquini <aquini@redhat.com>