[tools] tools include: Adopt kernel's refcount.h
Message-id: <20170830160540.26618-20-jolsa@redhat.com>
Patchwork-id: 183427
O-Subject: [PATCH RHEL7.5 BZ1480522 019/248] tools include: Adopt kernel's refcount.h
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 73a9bf95ed1c05698ecabe2f28c47aedfa61b52b
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Wed Feb 22 17:00:53 2017 -0300

description
===========
To aid in catching bugs when using atomics as a reference count.

This is a trimmed down version with just what is used by tools/ at
this point.

After this, the patches submitted by Elena for tools/ doing the
conversion from atomic_ to recount_ methods can be applied and tested.

To activate it, buint perf with:

  make DEBUG=1 -C tools/perf

Signed-off-by: Rafael Aquini <aquini@redhat.com>