[tools] perf probe: Factor out the ftrace README scanning
Message-id: <20170830160540.26618-52-jolsa@redhat.com>
Patchwork-id: 183467
O-Subject: [PATCH RHEL7.5 BZ1480522 051/248] perf probe: Factor out the ftrace README scanning
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 3da3ea7a8e205edc24b9491a459b46527c70b5b1
Author: Naveen N. Rao <naveen.n.rao@linux.vnet.ibm.com>
Date: Wed Mar 8 13:56:08 2017 +0530

description
===========
Simplify and separate out the ftrace README scanning logic into a
separate helper. This is used subsequently to scan for all patterns of
interest and to cache the result.

Since we are only interested in availability of probe argument type x,
we will only scan for that.

Signed-off-by: Rafael Aquini <aquini@redhat.com>