[tools] perf tools: Handle partial AUX records and print a warning
Message-id: <20170830160540.26618-63-jolsa@redhat.com>
Patchwork-id: 183475
O-Subject: [PATCH RHEL7.5 BZ1480522 062/248] perf tools: Handle partial AUX records and print a warning
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 05a1f47ed47a83736aca117aeee96e926cc0dfd0
Author: Alexander Shishkin <alexander.shishkin@linux.intel.com>
Date: Thu Mar 16 18:41:59 2017 +0200

description
===========
This patch decodes the 'partial' flag in AUX records and prints
a warning to the user, so that they don't have to guess why their
PT traces contain gaps (or missing altogether):

  Warning:
  AUX data had gaps in it 8 times out of 8!

  Are you running a KVM guest in the background?

Trying to be even more helpful, we will detect if the user's kvm driver sets up
exclusive VMX root mode for the entire lifespan of the kvm process:

  Reloading kvm_intel module with vmm_exclusive=0
  will reduce the gaps to only guest's timeslices.

Note however, that you'll still have gaps in cpu-wide traces even with
vmm_exclusive=0, but the number of gaps will be below 100% (as opposed to the
above example).

Currently this is the only reason for partial records.

Signed-off-by: Rafael Aquini <aquini@redhat.com>