[tools] perf report: Find the inline stack for a given address
Message-id: <20170830160540.26618-94-jolsa@redhat.com>
Patchwork-id: 183504
O-Subject: [PATCH RHEL7.5 BZ1480522 093/248] perf report: Find the inline stack for a given address
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit a64489c56c307bf0955f0489158c5ecf6aa10fe2
Author: Jin Yao <yao.jin@linux.intel.com>
Date: Sun Mar 26 04:34:26 2017 +0800

description
===========
It would be useful for perf to support a mode to query the inline stack
for a given callgraph address. This would simplify finding the right
code in code that does a lot of inlining.

The srcline.c has contained the code which supports to translate the
address to filename:line_nr. This patch just extends the function to let
it support getting the inline stacks.

It introduces the inline_list which will store the inline function
result (filename:line_nr and funcname).

If BFD lib is not supported, the result is only filename:line_nr.

Signed-off-by: Rafael Aquini <aquini@redhat.com>