[tools] perf annotate: Fix a bug following symbolic link of a build-id file
Message-id: <20170830160540.26618-99-jolsa@redhat.com>
Patchwork-id: 183510
O-Subject: [PATCH RHEL7.5 BZ1480522 098/248] perf annotate: Fix a bug following symbolic link of a build-id file
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 6ebd2547dd24daf95a21b2bc59931de8502afcc3
Author: Taeung Song <treeze.taeung@gmail.com>
Date: Mon Mar 27 16:10:36 2017 +0900

description
===========
It is wrong way to read link name from a build-id file.  Because a
build-id file is not anymore a symbolic link but build-id directory of
it is symbolic link, so fix it.

For example, if build-id file name gotten from
dso__build_id_filename() is as below,

  /root/.debug/.build-id/4f/75c7d197c951659d1c1b8b5fd49bcdf8f3f8b1/elf

To correctly read link name of build-id, use the build-id dir path that
is a symbolic link, instead of the above build-id file name like below.

  /root/.debug/.build-id/4f/75c7d197c951659d1c1b8b5fd49bcdf8f3f8b1

Fixes: 01412261d994 ("perf buildid-cache: Use path/to/bin/buildid/elf instead of path/to/bin/buildid")
Signed-off-by: Rafael Aquini <aquini@redhat.com>