[tools] perf test: Add a test case for SDT event
Message-id: <20170830160540.26618-103-jolsa@redhat.com>
Patchwork-id: 183515
O-Subject: [PATCH RHEL7.5 BZ1480522 102/248] perf test: Add a test case for SDT event
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 8e5dc848356ecf6ea8d27d641c4d7ad8d42fe92b
Author: Masami Hiramatsu <mhiramat@kernel.org>
Date: Tue Jul 12 19:06:05 2016 +0900

description
===========
Add a basic test case for SDT event support.  This test scans an SDT
event in perftools and check whether the SDT event is correctly stored
into the buildid cache.

Here is an example:

  ----
  $ perf test sdt -v
  47: Test SDT event probing                                   :
  --- start ---
  test child forked, pid 20732
  Found 72 SDTs in /home/mhiramat/ksrc/linux/tools/perf/perf
  Writing cache: %sdt_perf:test_target=test_target
  Cache committed: 0
  symbol:test_target file:(null) line:0 offset:0 return:0 lazy:(null)
  test child finished with 0
  ---- end ----
  Test SDT event probing: Ok
  ----

RHEL7
=====
This commit was omited in past for some reason,
but now we have dependencies on it, so there are
several context conflicts, nothing major.

Omitting the linux/filter.h include, because it adds
eBPF stuff which we don't have support for in RHEL7.

Signed-off-by: Rafael Aquini <aquini@redhat.com>