[tools] tools build: Fix feature detection redefinion of build flags
Message-id: <20170830160540.26618-151-jolsa@redhat.com>
Patchwork-id: 183570
O-Subject: [PATCH RHEL7.5 BZ1480522 150/248] tools build: Fix feature detection redefinion of build flags
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 9961aa665b70e47d6c80141c4a2482266010f246
Author: David Carrillo-Cisneros <davidcc@google.com>
Date: Tue Apr 11 23:49:14 2017 -0700

description
===========
This change is a follow up of https://lkml.org/lkml/2017/2/2/16

The patch above avoided redefining CC, CXX and PKG_CONFIG in feature
detection. The patch was not merged due to a unsolved concern with the
-MD flag.

Later, commit c8c188679ccf ("tools build: Use the same CC for feature
detection and actual build") did the change for CC and CXX but not
PKG_CONFIG.

This patch makes PKG_CONFIG consistent with CC and CXX and moves the -MD
to CFLAGS, as suggested by Jiri in the thread above.

Signed-off-by: Rafael Aquini <aquini@redhat.com>