[tools] perf str{filter, list}: Disentangle headers
Message-id: <20170830160540.26618-173-jolsa@redhat.com>
Patchwork-id: 183589
O-Subject: [PATCH RHEL7.5 BZ1480522 172/248] perf str{filter, list}: Disentangle headers
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit 8ec20b176c4be72d067fa18e33a4f156d1da9bc8
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Tue Apr 18 10:57:25 2017 -0300

description
===========
There are places where we just need a forward declaration, and others
were we need to include strlist.h and/or strfilter.h, reducing the
impact of changes in headers on the build time, do it.

RHEL7
=====
Skip BPF hunks, because it's not in RHEL7.

Signed-off-by: Rafael Aquini <aquini@redhat.com>