[tools] perf symbols: Kill dso__build_id_is_kmod()
Message-id: <20170830160540.26618-244-jolsa@redhat.com>
Patchwork-id: 183631
O-Subject: [PATCH RHEL7.5 BZ1480522 243/248] perf symbols: Kill dso__build_id_is_kmod()
Bugzilla: 1480522
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Arnaldo Carvalho de Melo <acme@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1480522
https://bugzilla.redhat.com/show_bug.cgi?id=1480522

upstream
========
commit b89fe63fbafe307fb72546f7a2320380bf41bdd4
Author: Namhyung Kim <namhyung@kernel.org>
Date: Thu Jun 8 16:31:09 2017 +0900

description
===========
The commit e7ee40475760 ("perf symbols: Fix symbols searching for module
in buildid-cache") added the function to check kernel modules reside in
the build-id cache.  This was because there's no way to identify a DSO
which is actually a kernel module.  So it searched linkname of the file
and find ".ko" suffix.

But this does not work for compressed kernel modules and now such DSOs
hCcave correct symtab_type now.  So no need to check it anymore.  This
patch essentially reverts the commit.

Signed-off-by: Rafael Aquini <aquini@redhat.com>