[scsi] qedf: Look at all descriptors when processing a clear virtual link
Message-id: <1504797133-30577-11-git-send-email-cdupuis@redhat.com>
Patchwork-id: 184115
O-Subject: [PATCH RHEL 7.5 e-stor 10/33] scsi: qedf: Look at all descriptors when processing a clear virtual link.
Bugzilla: 1461960
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chad Dupuis <chad.dupuis@cavium.com>

scsi commit 53c51adbe6f83eba94fd39000f98cf4235a89e4c

If there are multiple descriptors for a particular type in a clear virtual
link we receive, we will not process it correctly but rather take the last
value. This can cause us not to not flap the virtual link as the value from
the descriptors that we compare against the our stored FCF or fc_lport values
may not match.

Change this to do a comparison when processing the each descriptor instead of at
the end and then set a bool if we need to do the reset.

Signed-off-by: Chad Dupuis <chad.dupuis@cavium.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>