[scsi] qedf: Add bus_reset No-op
Message-id: <1504797133-30577-16-git-send-email-cdupuis@redhat.com>
Patchwork-id: 184144
O-Subject: [PATCH RHEL 7.5 e-stor 15/33] scsi: qedf: Add bus_reset No-op.
Bugzilla: 1461960
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chad Dupuis <chad.dupuis@cavium.com>

scsi commit b09fdc3aac6acc20587713e306692b12d31a2008

We need to add a bus reset no-op as without it some of the LUNs attached to a
vport may go offline when the error handler escalates to host reset due to not
having a bus reset handler in the driver. What happens is we escalate to host
reset which does a soft link down/link up to reset the adapter. However with
multiple vports attached it's been observed that if the vports do log back into
the target within 5 seconds, the SCSI layer offlines the devices most likely
due to a TUR timing out to verify that the device is online. Adding a bus
reset handler will cause the TUR to be sent after the bus reset handler where
the devices will still be online if the bus reset is initiated by sg_reset
(which is the case in the test that was failing). The bus reset will succeed
and not needlessly bring the device offline/online.

Signed-off-by: Chad Dupuis <chad.dupuis@cavium.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>