[scsi] qedf: Use granted MAC from the FCF for the FCoE source address if it is available
Message-id: <1504797133-30577-29-git-send-email-cdupuis@redhat.com>
Patchwork-id: 184151
O-Subject: [PATCH RHEL 7.5 e-stor 28/33] scsi: qedf: Use granted MAC from the FCF for the FCoE source address if it is available.
Bugzilla: 1461960
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Chad Dupuis <chad.dupuis@cavium.com>

scsi commit a3cd42a9d627fd7c472664e454108829886e123f

Currently in the driver we've been using the fc_fcoe_set_mac() function to
set the source MAC for FCoE traffic.  This works well in most cases as it
uses the spec. default FCF-MAC.  However, if the administrator changes the
FCF-MAC switch, then any FCoE traffic we send will be dropped by the
switch.

Instead we should check the granted MAC from the FLOGI payload and use that
address if it is present.  Otherwise, fall back to using the the default
FCF-MAC and the fabric ID of the port as the FCoE MAC address.

Once this address is known we need to set it when doing non-offload
traffic, offload traffic and setting the data_src_address libfcoe uses for
FIP keep alive messages.

Signed-off-by: Chad Dupuis <chad.dupuis@cavium.com>
Signed-off-by: Martin K. Petersen <martin.petersen@oracle.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>