[kernel] cpu/hotplug: Disable prove_locking for cpu_hotplug.mutex
Message-id: <1504800156-11756-3-git-send-email-longman@redhat.com>
Patchwork-id: 184130
O-Subject: [RHEL7.5 PATCH 2/3] cpu/hotplug: Disable prove_locking for cpu_hotplug.mutex
Bugzilla: 1488956
RH-Acked-by: Larry Woodman <lwoodman@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1488956

This is a RHEL7 specific patch.

The cpu_hotplug structure has 2 separate lockdep allocation debug
structures. One is the core dep_map in the structure itself and
the other one is in the cpu_hotplug.lock mutex. The core dep_map is
tracked as a rwlock with get_online_cpus() treated as a read lock
and cpu_hotplug_begin() as a write lock.

The try_get_online_cpus() can work around potential recursive lock
taking issue reported by lockdep. However, the lockdep structure in
the mutex will still report lockdep error because of the mutex_lock()
in put_online_cpus(). It is actually not a real problem as the success
of try_get_online_cpus() means other cpus cannot start a hotplug event
sequence before put_onlines_cpus(). Other mutex lock holders can only
increment and decrement the reference count which is pretty quick.

In fact, the lockdep warning is about a CPU taking the write lock
in the hotplug event. An interrupt happens and eventually a softirq
code running the shrinker which tries to take a read lock leading to
a deadlock. That won't happen anymore as the trylock will fail and
the alternative RCU path will be taken.

To avoid false positive of this kind, the locking correctness tracking
in the mutex itself is turned off. Locking correctness tracking will
still be done in the core dep_map structure as a rwlock.

Signed-off-by: Waiman Long <longman@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>