[netdrv] reference new cavium directory in Makefile and Kconfig
Message-id: <20170907192214.GA20480@felix-thinkpad.cavium.com>
Patchwork-id: 184155
O-Subject: [PATCH v2 RHEL7.5 003/202] ethernet: reference new cavium directory in Makefile and Kconfig
Bugzilla: 1451436
RH-Acked-by: Paolo Abeni <pabeni@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1451436

Upstream Status: linux.git
commit 4863dea3fab01734768c9f7fc2c1590a8f1f6266

Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
Only a small portion of the upstream commit is relevant to the BZ; the
majority is irrelevant because the commit pertains to the Cavium ThunderX
network controller--which is not liquidio.  ThunderX made it to upstream a
couple of weeks ahead of liquidio.

Here are the tiny relevant parts of the upstream commit that are in this
patch:

    drivers/net/ethernet/Kconfig
    drivers/net/ethernet/Makefile

One line was added in each of the files above to reference the new cavium
directory which contains the liquidio subdirectory (in upstream, that
directory also contains the thunder subdirectory).

To avoid confusion, I'm omitting the upstream commit log.  It doesn't
mention liquidio; the text is not suitable for this patch.

Signed-off-by: Felix Manlunas <fmanluna@redhat.com>