[netdrv] liquidio: fix build errors when CONFIG_DYNAMIC_DEBUG=y and both PF and VF drivers are built as loadable modules
Message-id: <20170907194425.GA21871@felix-thinkpad.cavium.com>
Patchwork-id: 184355
O-Subject: [PATCH v2 RHEL7.5 200/202] liquidio: fix build errors when CONFIG_DYNAMIC_DEBUG=y and both PF and VF drivers are built as loadable modules
Bugzilla: 1451436
RH-Acked-by: Paolo Abeni <pabeni@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1451436

Upstream Status: none, this is RHEL only

The build errors described below do not happen in upstream.  They happen in
RHEL7.

When the following are all present in the kernel .config file:

    CONFIG_DYNAMIC_DEBUG=y
    CONFIG_LIQUIDIO=m
    CONFIG_LIQUIDIO_VF=m

these build errors happen:

      CC [M]  drivers/net/ethernet/cavium/liquidio/lio_ethtool.o
    In file included from include/linux/printk.h:241:0,
                     from include/linux/kernel.h:13,
                     from include/linux/plist.h:78,
                     from include/linux/pm_qos.h:7,
                     from include/linux/netdevice.h:28,
                     from drivers/net/ethernet/cavium/liquidio/lio_ethtool.c:18:
    drivers/net/ethernet/cavium/liquidio/octeon_main.h: In function `octeon_unmap_pci_barx':
    include/linux/dynamic_debug.h:66:14: error: `KBUILD_MODNAME' undeclared (first use in this function)
       .modname = KBUILD_MODNAME,   \
              ^
    include/linux/dynamic_debug.h:84:2: note: in expansion of macro `DEFINE_DYNAMIC_DEBUG_METADATA'
      DEFINE_DYNAMIC_DEBUG_METADATA(descriptor, fmt);  \

and so on ...

The reason is:  Some .c files are common to liquidio.ko and liquidio_vf.ko.
When such .c files are compiled (with both CONFIG_LIQUIDIO=m and
CONFIG_LIQUIDIO_VF=m), the KBUILD_MODNAME macro definition is omitted in
the gcc command line (not the case in upstream where it's always defined
via gcc -D option).  But dynamic debug requires that the KBUILD_MODNAME
macro be defined.

Fix it by conditionally #defining (in a common header file that's #included
by the affected .c files) KBUILD_MODNAME to have the same value as
KBUILD_BASENAME; the condition being:  KBUILD_MODNAME was not previously
defined.

This sort of mimics upstream in terms of defining KBUILD_MODNAME to have
the proper value which happens to be the same as that of KBUILD_BASENAME
when compiling a .c file that's common to at least two loadable modules.
It's just that in upstream, KBUILD_MODNAME is always defined in the gcc
command line.

By the way in upstream and in RHEL7, KBUILD_BASENAME always gets defined in
the gcc command line on a per .c file basis.

Signed-off-by: Felix Manlunas <fmanluna@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>