[scsi] megaraid: remove DRIVER_ATTR() usage
Message-id: <20170908141140.6631-5-thenzl@redhat.com>
Patchwork-id: 184375
O-Subject: [RHEL7.5 e-stor PATCH 04/19] scsi: remove DRIVER_ATTR() usage
Bugzilla: 1458133
RH-Acked-by: Chris Leech <cleech@redhat.com>
RH-Acked-by: Maurizio Lombardi <mlombard@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

It's better to use the DRIVER_ATTR_RW() and DRIVER_ATTR_RO() macros to
explicitly show that this is a read/write or read/only sysfs file.  So
convert the remaining SCSI drivers that use the old style to use the
newer macros.

Bonus is that this removes some checkpatch.pl warnings :)

This is part of a series to drop DRIVER_ATTR() from the tree entirely.

commit e14a3967723ed4a1dacd180cfc560ffc510ea2a6
(limited to megaraid_sas)

Cc: "James E.J. Bottomley" <jejb@linux.vnet.ibm.com>
Cc: "Martin K. Petersen" <martin.petersen@oracle.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>