[net] l2cap: prevent stack overflow on incoming bluetooth packet
Message-id: <20170908184119.14875-1-nhorman@redhat.com>
Patchwork-id: 184404
O-Subject: [RHEL7.5 PATCH] l2cap: prevent stack overflow on incoming bluetooth packet
Bugzilla: 1489789
CVE: CVE-2017-1000251
RH-Acked-by: Jiri Benc <jbenc@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Neil Horman <nhorman@redhat.com>

This came in as a security vulnerability today, and is expected to be
disclosed on Tuesday, Sept 12.  If a bluetooth packet arrives, the
function l2cap_parse_conf_rsp parses the config elements in the packets,
which can be of arbitrary length.  Since the function does not receive
the maximum length of the output buffer (which is on the stack), a
malformed packet can overwrite the stack, and lead to arbitrary code
execution.  Further, given that these packets do not require an
established connection to be parsed, simpy having bluetooth enabled on a
given system may make a host vulnerable.

This patch was provided by the research firm that informed us of the
issue, but has not gone upstream yet (it will on Tuesday, the disclosure
date).  Until then, we need to review this patch, and get it integrated
asap so that we are no longer vulnerable to the attack on the day the
CVE is disclosed.

Resolves bz 1489789

BZ: 1489789
Brew: https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=14025221
Testing: Currently in test in parallel with this posting, due to the
time sensitive nature of the bug
Upstream Status: RHEL only until tuesday, when the CVE is disclosed.  We
will need to track the upstream progression of this bug, and adjust the
patch, should its content change between now and then.

Signed-off-by: Rafael Aquini <aquini@redhat.com>