[mm] page_cgroup: Fix Kernel bug during boot with memory cgroups enabled
Message-id: <1505746434-712-1-git-send-email-lwoodman@redhat.com>
Patchwork-id: 185491
O-Subject: [RHEL7.5 PATCH V2] Fix Kernel bug during boot with memory cgroups enabled on Dell R7415
Bugzilla: 1483747
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

RHEL-only patch, not upstream

There is a bug in the kernel function page_cgroup_init() that assumes
a section starts in another NUMA node and therefore doesn't allocate
page_cgroup memory for that section.

page_cgroup_init() is supposed to loop through each node and in each
node it will go through each section (128MB, as defined in the kernel),
allocate space for the page_cgroup structures that are used to keep track
of which cgroup various pages belong to.  At the beginning of each 128MB
section, it checks to see if the starting address of that section is in
the NUMA node that it is currently setting up.  If not, it assumes that
the section starts in another NUMA node, and doesn't allocate page_cgroup
memory for that section.  If the memory section belongs to a lower node,
it would have already been initialized and calling init_section_page_cgroup()
again won't hurt anything.  The first thing that function does is check
if that section's page_cgroups were already initialized and if so it
will immediately return.

This patch will fix this and was upstream but is now gone since
the memory cgroup init stuff has been rewriten.

Signed-off-by: Rafael Aquini <aquini@redhat.com>