[mm] fix deadlock when using dm-thin on loopback device
Message-id: <alpine.LRH.2.02.1709190909180.13320@file01.intranet.prod.int.rdu2.redhat.com>
Patchwork-id: 185883
O-Subject: [RHEL7.5 PATCH] bz1382654: fix deadlock when using dm-thin on loopback device
Bugzilla: 1382654
RH-Acked-by: Mike Snitzer <snitzer@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>

This patch fixes deadlocks when using dm-thin on loopback device.

The reason for the deadlocks is this:

* dm-thin uses dm-bufio to perform metadata I/O
* dm-bufio holds a lock and submits a bio to the underlying loopback device
* the bio is redirected to the filesystem (xfs in the test case)
* the VFS layer performs some memory allocation with GFP_KERNEL flags
* the allocation eventually reaches the function shrink_slab to shrink memory
* shrink_slab calls the shrink function in dm-bufio
* the shrink function attempts to take a mutex that is already held

([<000000000069593c>] __schedule+0x53c/0xbb8)
 [<000000000069649c>] schedule_preempt_disabled+0x3c/0x48
 [<0000000000693f28>] __mutex_lock_slowpath+0xe0/0x228
 [<00000000006940de>] mutex_lock+0x6e/0x78
 [<000003ff8002d3f2>] shrink+0x62/0x1c8 [dm_bufio]
 [<0000000000266e58>] shrink_slab+0xd8/0x3f8
 [<000000000026a5dc>] do_try_to_free_pages+0x4d4/0x658
 [<000000000026a83c>] try_to_free_pages+0xdc/0x198
 [<000000000025b74e>] __alloc_pages_nodemask+0x856/0xbd0
 [<00000000002b2492>] new_slab+0x9a/0x368
 [<00000000002b4faa>] ___slab_alloc.constprop.57+0x2b2/0x3a0
 [<000000000068dd9e>] __slab_alloc.constprop.56+0x46/0x70
 [<00000000002b5830>] kmem_cache_alloc+0x1c0/0x228
 [<000000000042d188>] __radix_tree_preload+0x48/0xd0
 [<000000000025072c>] __add_to_page_cache_locked+0x8c/0x290
 [<00000000002509d6>] add_to_page_cache_lru+0x56/0xf0
 [<0000000000250efe>] grab_cache_page_write_begin+0xd6/0x118
 [<000003ff804401c6>] xfs_vm_write_begin+0x56/0x108 [xfs]
 [<00000000002504f6>] generic_file_buffered_write+0x13e/0x2e8
 [<000003ff80450994>] xfs_file_buffered_aio_write+0x16c/0x228 [xfs]
 [<000003ff80450c30>] xfs_file_aio_write+0x1e0/0x1f0 [xfs]
 [<00000000002d28a4>] do_sync_write+0x9c/0xd8
 [<000003ff809f4ed0>] __do_lo_send_write+0xb8/0x180 [loop]
 [<000003ff809f5006>] do_lo_send_direct_write+0x6e/0x88 [loop]
 [<000003ff809f53ae>] do_bio_filebacked+0x28e/0x338 [loop]
 [<000003ff809f554e>] loop_thread+0xf6/0x2b0 [loop]
 [<00000000001728d2>] kthread+0xea/0xf0
 [<0000000000697c8a>] kernel_thread_starter+0x6/0xc
 [<0000000000697c84>] kernel_thread_starter+0x0/0xc

In order to fix the bug, we must make sure that all page cache allocations
for loopback-backed device are done with the flag GFP_NOIO. The kernel
already maintains allocation flags for struct address_space (the flags
variable, it can be retrieved with mapping_gfp_mask()) and when the file
is used as loopback backing file, it is already set to GFP_NOIO.

The problem is that these flags are used inconsistently, i.e. even if
mapping->flags is GFP_NOIO, the kernel sometimes does GFP_KERNEL
allocations (and these allocations cause the deadlock). We must fix it, so
that all allocations for a given file are done with mapping->flags.

This bug was already fixed in the upstream kernel - the patches
6afdb859b710 ("mm: do not ignore mapping_gfp_mask in page cache allocation
paths") and 063d99b4fa76 ("mm, fs: obey gfp_mapping for
add_to_page_cache()"). The upstream code diverged significantly from the
RHEL7 code, so the patches are different. (I choose to restrict gfp_mask
in __add_to_page_cache_locked instead of all the callers of this functions
like in the upstream patch)

Bruno Goncalves tested the patch on a s390x machine where he encountered
the deadlocks - https://bugzilla.redhat.com/show_bug.cgi?id=1382654#c29

Signed-off-by: Mikulas Patocka <mpatocka@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>