[sound] alsa: timer: Wrap with spinlock for queue access
Message-id: <150608649521-8-patchset-jkysela@redhat.com>
Patchwork-id: 186294
O-Subject: [RHEL 7.5 PATCH v3 07/10] ALSA: timer: Wrap with spinlock for queue access
Bugzilla: 1465999
CVE: CVE-2017-1000380
RH-Acked-by: John Linville <linville@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

For accessing the snd_timer_user queue indices, we take tu->qlock.
But it's forgotten in a couple of places.

The one in snd_timer_user_params() should be safe without the
spinlock as the timer is already stopped.  But it's better for
consistency.

The one in poll is just a read-out, so it's not inevitably needed, but
it'd be good to make the result consistent, too.

Tested-by: Alexander Potapenko <glider@google.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>

Signed-off-by: Jaroslav Kysela <jkysela@redhat.com>
(cherry picked from commit d7f910bfedd863d13ea320030fe98e42d0938ed5)
Bugzilla: 1465999
Signed-off-by: Rafael Aquini <aquini@redhat.com>