[x86] intel_rdt/cqm: Add RMID (Resource monitoring ID) management
Message-id: <20170925122043.19483-11-jolsa@redhat.com>
Patchwork-id: 186426
O-Subject: [PATCH RHEL7.5 10/42] x86/intel_rdt/cqm: Add RMID (Resource monitoring ID) management
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1457533
https://bugzilla.redhat.com/show_bug.cgi?id=1457533

upstream
========
commit edf6fa1c4a951b3a03e94b63e6483c5d9da3ab11
Author: Vikas Shivappa <vikas.shivappa@linux.intel.com>
Date: Tue Jul 25 14:14:28 2017 -0700

description
===========
Hardware uses RMID(Resource monitoring ID) to keep track of each of the
RDT events associated with tasks. The number of RMIDs is dependent on
the SKU and is enumerated via CPUID. We add support to manage the RMIDs
which include managing the RMID allocation and reading LLC occupancy
for an RMID.

RMID allocation is managed by keeping a free list which is initialized
to all available RMIDs except for RMID 0 which is always reserved for
root group. RMIDs goto a limbo list once they are
freed since the RMIDs are still tagged to cache lines of the tasks which
were using them - thereby still having some occupancy. They continue to
be in limbo list until the occupancy < threshold_occupancy. The
threshold_occupancy is a user configurable value.
OS uses IA32_QM_CTR MSR to read the occupancy associated with an RMID
after programming the IA32_EVENTSEL MSR with the RMID.

[Tony: Improved limbo search]

Signed-off-by: Rafael Aquini <aquini@redhat.com>