[x86] intel_rdt: Simplify info and base file lists
Message-id: <20170925122043.19483-12-jolsa@redhat.com>
Patchwork-id: 186428
O-Subject: [PATCH RHEL7.5 11/42] x86/intel_rdt: Simplify info and base file lists
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1457533
https://bugzilla.redhat.com/show_bug.cgi?id=1457533

upstream
========
commit 5dc1d5c6bac2cfe3420cf353dfb0ef2e543f7c10
Author: Tony luck <tony.luck@intel.com>
Date: Tue Jul 25 14:14:29 2017 -0700

description
===========
The info directory files and base files need to be different for each
resource like cache and Memory bandwidth. With in each resource, the
files would be further different for monitoring and ctrl. This leads to
a lot of different static array declarations given that we are adding
resctrl monitoring.

Simplify this to one common list of files and then declare a set of
flags to choose the files based on the resource, whether it is info or
base and if it is control type file. This is as a preparation to include
monitoring based info and base files.

No functional change.

[Vikas: Extended the flags to have few bits per category like resource,
	info/base etc]

Signed-off-by: Rafael Aquini <aquini@redhat.com>