[x86] intel_rdt/cqm: Add cpus file support
Message-id: <20170925122043.19483-19-jolsa@redhat.com>
Patchwork-id: 186390
O-Subject: [PATCH RHEL7.5 18/42] x86/intel_rdt/cqm: Add cpus file support
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1457533
https://bugzilla.redhat.com/show_bug.cgi?id=1457533

upstream
========
commit a9fcf8627dc01049c390023bbb0323db3c785b91
Author: Vikas Shivappa <vikas.shivappa@linux.intel.com>
Date: Tue Jul 25 14:14:36 2017 -0700

description
===========
The cpus file is extended to support resource monitoring. This is used
to over-ride the RMID of the default group when running on specific
CPUs. It works similar to the resource control. The "cpus" and
"cpus_list" file is present in default group, ctrl_mon groups and
monitor groups.

Each "cpus" file or cpu_list file reads a cpumask or list showing which
CPUs belong to the resource group. By default all online cpus belong to
the default root group. A CPU can be present in one "ctrl_mon" and one
"monitor" group simultaneously. They can be added to a resource group by
writing the CPU to the file. When a CPU is added to a ctrl_mon group it
is automatically removed from the previous ctrl_mon group. A CPU can be
added to a monitor group only if it is present in the parent ctrl_mon
group and when a CPU is added to a monitor group, it is automatically
removed from the previous monitor group. When CPUs go offline, they are
automatically removed from the ctrl_mon and monitor groups.

Signed-off-by: Rafael Aquini <aquini@redhat.com>