[x86] intel_rdt/cqm: Add rmdir support
Message-id: <20170925122043.19483-23-jolsa@redhat.com>
Patchwork-id: 186391
O-Subject: [PATCH RHEL7.5 22/42] x86/intel_rdt/cqm: Add rmdir support
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1457533
https://bugzilla.redhat.com/show_bug.cgi?id=1457533

upstream
========
commit f3cbeacaa06e2b8c2f3ce8531e9aa3fe1f2762cd
Author: Vikas Shivappa <vikas.shivappa@linux.intel.com>
Date: Tue Jul 25 14:14:40 2017 -0700

description
===========
Resource groups (ctrl_mon and monitor groups) are represented by
directories in resctrl fs. Add support to remove the directories.

When a ctrl_mon directory is removed all the cpus and tasks are assigned
back to the root rdtgroup. When a monitor group is removed the cpus and
tasks are returned to the parent control group.

Signed-off-by: Rafael Aquini <aquini@redhat.com>