[x86] intel_rdt: Show bitmask of shareable resource with other executing units
Message-id: <20170925122043.19483-34-jolsa@redhat.com>
Patchwork-id: 186417
O-Subject: [PATCH RHEL7.5 33/42] x86/intel_rdt: Show bitmask of shareable resource with other executing units
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1457533
https://bugzilla.redhat.com/show_bug.cgi?id=1457533

upstream
========
commit 0dd2d7494cd818d06a2ae1cd840cd62124a2d25e
Author: Fenghua Yu <fenghua.yu@intel.com>
Date: Tue Jul 25 15:39:04 2017 -0700

description
===========
CPUID.(EAX=0x10, ECX=res#):EBX[31:0] reports a bit mask for a resource.
Each set bit within the length of the CBM indicates the corresponding
unit of the resource allocation may be used by other entities in the
platform (e.g. an integrated graphics engine or hardware units outside
the processor core and have direct access to the resource). Each
cleared bit within the length of the CBM indicates the corresponding
allocation unit can be configured to implement a priority-based
allocation scheme without interference with other hardware agents in
the system. Bits outside the length of the CBM are reserved.

More details on the bit mask are described in x86 Software Developer's
Manual.

The bitmask is shown in "info" directory for each resource. It's
up to user to decide how to use the bitmask within a CBM in a partition
to share or isolate a resource with other executing units.

Signed-off-by: Rafael Aquini <aquini@redhat.com>