[x86] intel_rdt/mbm: Fix MBM overflow handler during CPU hotplug
Message-id: <20170925122043.19483-37-jolsa@redhat.com>
Patchwork-id: 186413
O-Subject: [PATCH RHEL7.5 36/42] x86/intel_rdt/mbm: Fix MBM overflow handler during CPU hotplug
Bugzilla: 1457533
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1457533
https://bugzilla.redhat.com/show_bug.cgi?id=1457533

upstream
========
commit bbc4615e0b7df5e21d0991adb4b2798508354924
Author: Vikas Shivappa <vikas.shivappa@linux.intel.com>
Date: Tue Aug 15 18:00:42 2017 -0700

description
===========
When a CPU is dying, the overflow worker is canceled and rescheduled on a
different CPU in the same domain. But if the timer is already about to
expire this essentially doubles the interval which might result in a non
detected overflow.

Cancel the overflow worker and reschedule it immediately on a different CPU
in same domain. The work could be flushed as well, but that would
reschedule it on the same CPU.

[ tglx: Rewrote changelog once again ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>