[kernel] tracing/kprobes: Allow to create probe with a module name starting with a digit
Message-id: <20170926125631.22703-3-jolsa@redhat.com>
Patchwork-id: 186618
O-Subject: [PATCH RHEL7.5 2/2] tracing/kprobes: Allow to create probe with a module name starting with a digit
Bugzilla: 1422117
RH-Acked-by: Oleg Nesterov <oleg@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1422117
https://bugzilla.redhat.com/show_bug.cgi?id=1422117

upstream
========
commit 9e52b32567126fe146f198971364f68d3bc5233f
Author: Sabrina Dubroca <sd@queasysnail.net>
Date:   Thu Jun 22 11:24:42 2017 +0200

description
===========
Always try to parse an address, since kstrtoul() will safely fail when
given a symbol as input. If that fails (which will be the case for a
symbol), try to parse a symbol instead.

This allows creating a probe such as:

    p:probe/vlan_gro_receive 8021q:vlan_gro_receive+0

Which is necessary for this command to work:

    perf probe -m 8021q -a vlan_gro_receive

Signed-off-by: Rafael Aquini <aquini@redhat.com>