[kernel] radix-tree: RHEL-only kABI patch
Message-id: <1506453762-7009-1-git-send-email-lwoodman@redhat.com>
Patchwork-id: 186637
O-Subject: [RHEL7.5 PATCH V2] radix-tree: RHEL-only kABI patch
Bugzilla: 1383492
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

V2 adds more comments and new kabi macro RH_KABI_REPLACE2() to
include/linux/rh_kabi.h and replaces the "ifdef __GENKSYMS__"
in the radix_tree_node struct with that new macro.  This is an
attempt to eliminate the use of GENKSYMS.
--------------------------------------------------------------

The radix tree functions are (surprisingly) not covered by our kabi
guarantees.  However, because you can reach the radix_tree_root and
radix_tree_node structures through various other data structures that
are covered by kabi, the checker complained.

The radix_tree_root is embedded in other data structures.  So, for
that one, we have to maintain the same size.  RH_KABI_DEPRECATE takes
care of this.  The radix_tree_node is only pointed to (not embedded),
so we use the new/included RH_KABI_REPLACE2() macro to satisfy the kabi
check.  Because we don't need to preserve the size of the structure,
there was no need to play games with anonymous unions or structs.

NOTE: this patch set *will* break any module that makes use of the
radix tree library.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>