[x86] x86/mce: Include the PPIN in MCE records when available
Message-id: <20170928041234.6006-2-swood@redhat.com>
Patchwork-id: 186934
O-Subject: [RHEL7.5 BZ 1440834 1/2] x86/mce: Include the PPIN in MCE records when available
Bugzilla: 1440834
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Steve Best <sbest@redhat.com>

Intel Xeons from Ivy Bridge onwards support a processor identification
number set in the factory. To the user this is a handy unique number to
identify a particular CPU. Intel can decode this to the fab/production
run to track errors. On systems that have it, include it in the machine
check record. I'm told that this would be helpful for users that run
large data centers with multi-socket servers to keep track of which CPUs
are seeing errors.

Boris:
* Add some clarifying comments and spacing.
* Mask out [63:2] in the disabled-but-not-locked case
* Call the MSR variable "val" for more readability.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Cc: Ashok Raj <ashok.raj@intel.com>
Cc: linux-edac <linux-edac@vger.kernel.org>
Cc: x86-ml <x86@kernel.org>
Link: http://lkml.kernel.org/r/20161123114855.njguoaygp3qnbkia@pd.tnic
Signed-off-by: Borislav Petkov <bp@suse.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
(cherry picked from commit 3f5a7896a5096fd50030a04d4c3f28a7441e30a5)
Signed-off-by: Scott Wood <swood@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	arch/x86/include/asm/cpufeatures.h