[fs] revert "cifs: Fix null pointer deref during read resp processing"
Message-id: <20171002230614.31788-1-lsahlber@redhat.com>
Patchwork-id: 187360
O-Subject: [RHEL7.5 BZ 1403319 fs] Revert "[fs] cifs: Fix null pointer deref during read resp processing"
Bugzilla: 1403319
RH-Acked-by: Benjamin Coddington <bcodding@redhat.com>
RH-Acked-by: Sachin Prabhu <sprabhu@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Bugzilla: 1403319
Build: 14159546

This reverts commit c73646848845e69bca8ce1f49f78dd5421da07ae.

This patch was originally part of the SMB3 encryption patch series
that just went in.
The backport only containes those parts needed for the SMB3 encryption series.

As part of 1403319 we need the full set of changes in the original
patch why we need to revert this patch as we will bring the full
patch in via that bug.

This patch needs to be applied before the other two patches in 1403319
can be applied.

Signed-off-by: Rafael Aquini <aquini@redhat.com>