[nvme] rdma: register ib_client to not deadlock in device removal
Message-id: <1507192397-86938-167-git-send-email-dmilburn@redhat.com>
Patchwork-id: 189225
O-Subject: [RHEL7.5 PATCH v3 166/206] nvmet-rdma: register ib_client to not deadlock in device removal
Bugzilla: 1457880 1456486 1454365
RH-Acked-by: Ewan Milne <emilne@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Sagi Grimberg <sagi@grimberg.me>

We can deadlock in case we got to a device removal
event on a queue which is already in the process of
destroying the cm_id is this is blocking until all
events on this cm_id will drain. On the other hand
we cannot guarantee that rdma_destroy_id was invoked
as we only have indication that the queue disconnect
flow has been queued (the queue state is updated before
the realease work has been queued).

So, we leave all the queue removal to a separate ib_client
to avoid this deadlock as ib_client device removal is in
a different context than the cm_id itself.

Reported-by: Shiraz Saleem <shiraz.saleem@intel.com>
Tested-by: Shiraz Saleem <shiraz.saleem@intel.com>
Signed-off-by: Sagi Grimberg <sagi@grimberg.me>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
(cherry picked from commit f1d4ef7d88832444e8dfeb0e85e19d3b6ecb5011)

BZ 1454365
BZ 1456486
BZ 1457880

Signed-off-by: David Milburn <dmilburn@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>